requirements.txt
pandas==2.2.1
numpy==1.26.0
matplotlib.pyplot==3.5.3
seaborn==0.13.2
scikit-learn==1.4.2
